backend/pyproject.toml
[project]
name = "grpc-attendance"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "beanie>=2.0.0",
    "black>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqladmin>=0.21.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.74.0",
]
lint = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
]

[tool.black]
line-length = 69
